requirements.txt
# All
coverage

# Crawlers
scrapy==2.5.0
splash
scrapy-splash
scrapyjs
selenium
scrapy-selenium
avro-python3

webdriver-manager

# Scheduler
celery==4.4.7
django-celery-results
django-celery-beat
flower==0.9.5
vine

# Django API
django
djangorestframework==3.12.*
django-cors-headers
django-environ==0.4.*
psycopg2-binary>=2.8
pycodestyle
coverage
dnspython
drf-nested-routers
